refactor: switch relative imports to absolute to avoid namespace conflicts
